Move validation evidence tests into webcodex-core
