+++ crates/webcodex-core/src/validation_evidence_tests.rs
@@ -1,5 +1,4 @@
-use crate::tool_runtime::cargo::parse_cargo_test_counts;
-use crate::tool_runtime::validation_parser::{
+use crate::validation_evidence::{
     parse_cargo_check_diagnostics, parse_cargo_test_diagnostics, parse_go_test_diagnostics,
     NO_STABLE_DIAGNOSTICS_REASON, PARSER_KIND, PARSER_LIMITATIONS, PARSER_VERSION,
 };
@@ -652,27 +651,6 @@ fn cargo_test_parser_aggregated_truncation_uses_summed_failed_count() {
     let summary = diagnostics.test_summary.as_ref().unwrap();
     assert_eq!(summary.passed, Some(1));
     assert_eq!(summary.failed, Some(5));
-}
-
-#[test]
-fn parse_cargo_test_counts_aggregates_multiple_harness_summaries() {
-    let (passed, failed) = parse_cargo_test_counts(
-        "test result: ok. 2 passed; 0 failed; 1 ignored\n\
-         test result: FAILED. 3 passed; 1 failed; 0 ignored\n\
-         test result: ok. 0 passed; 0 failed; 2 ignored\n",
-    );
-    assert_eq!(passed, Some(5));
-    assert_eq!(failed, Some(1));
-}
-
-#[test]
-fn parse_cargo_test_counts_does_not_use_last_summary_wins() {
-    let (passed, failed) = parse_cargo_test_counts(
-        "test result: FAILED. 10 passed; 4 failed; 0 ignored; 0 measured; 0 filtered out\n\
-         test result: ok. 1 passed; 0 failed; 0 ignored; 0 measured; 0 filtered out\n",
-    );
-    assert_eq!(passed, Some(11));
-    assert_eq!(failed, Some(4));
 }
 
 #[test]
